refactor: Move component management related code to separate store
